test: split card-dag tests out of test_graph_inference
Bundle the three card-related tests into test_card_dag.py:
- test_transform_flow_graph_supports_explicit_endpoints
- test_transform_flow_graph_keeps_legacy_start_end_detection
- test_default_card_includes_custom_graph_endpoints

All three target the cards' graph data layer (transform_flow_graph
and the DAG component render path). Keeping them next to the
graph-execution tests in test_graph_inference.py blurred file
boundaries and meant test_graph_inference.py had to import
DefaultCardJSON, transform_flow_graph, and a four-line importlib
helper just for the one render test.

To support the render test cleanly, hoist the FlowSpec class loader
into conftest.py as _load_flow_class plus a custom_named_card_flow
fixture. DefaultCardJSON.render() does getattr(self.flow, step_name)
and inspect.getsource(...) to build the Task Code panel and crashes
if flow is None, so the test still needs the actual class object;
moving the loader into conftest gets the importlib dance out of the
test files.

No behavior change.
